feat: Update Button theme
