+++ .eslintrc.json
@@ -7,7 +7,7 @@
     "@typescript-eslint/no-shadow": "off",
     "react-hooks/exhaustive-deps": "off",
     "linebreak-style": "off",
-    "indent": "off"
+    "indent": "off",
     "no-console": "off",
     "no-restricted-syntax": 0
   }
